+++ src/lib.rs
@@ -12,27 +12,29 @@ pub mod render;
 pub mod window;
 
 use std::collections::BTreeMap;
+use std::fmt::{self, Display, Formatter};
 use std::num::NonZeroUsize;
 use std::sync::Arc;
 use std::sync::atomic::{self, AtomicBool};
 use std::thread::available_parallelism;
 
-use interface::Interface;
-use vulkano::VulkanLibrary;
-use vulkano::device::physical::{PhysicalDevice, PhysicalDeviceType};
-use vulkano::device::{
-    self, Device, DeviceCreateInfo, DeviceExtensions, DeviceFeatures, QueueCreateInfo, QueueFlags,
-};
-use vulkano::instance::{
-    Instance, InstanceCreateFlags, InstanceCreateInfo, InstanceExtensions, Version,
-};
-
 mod vk {
+    pub use vulkano::device::physical::{PhysicalDevice, PhysicalDeviceType};
+    pub use vulkano::device::{
+        Device, DeviceCreateInfo, DeviceExtensions, DeviceFeatures, Queue, QueueCreateInfo,
+        QueueFlags,
+    };
+    pub use vulkano::instance::{
+        Instance, InstanceCreateFlags, InstanceCreateInfo, InstanceExtensions, Version,
+    };
+    pub use vulkano::library::LoadingError;
+    pub use vulkano::{Validated, VulkanError, VulkanLibrary};
     pub use vulkano_taskgraph::resource::Resources;
 }
 
 use crate::image::ImageCache;
 use crate::input::Input;
+use crate::interface::Interface;
 use crate::interval::Interval;
 use crate::render::{MSAA, VSync};
 use crate::window::WindowManager;
@@ -40,16 +42,16 @@ use crate::window::WindowManager;
 /// Options for Basalt's creation and operation.
 pub struct BasaltOptions {
     // Instance Options
-    require_instance_extensions: InstanceExtensions,
-    prefer_instance_extensions: InstanceExtensions,
+    require_instance_extensions: vk::InstanceExtensions,
+    prefer_instance_extensions: vk::InstanceExtensions,
     // Physical Device Selection
     portability_subset: bool,
     prefer_integrated_gpu: bool,
     // Device Options
-    require_device_extensions: DeviceExtensions,
-    prefer_device_extensions: DeviceExtensions,
-    require_device_features: DeviceFeatures,
-    prefer_device_features: DeviceFeatures,
+    require_device_extensions: vk::DeviceExtensions,
+    prefer_device_extensions: vk::DeviceExtensions,
+    require_device_features: vk::DeviceFeatures,
+    prefer_device_features: vk::DeviceFeatures,
     // Window Options
     winit_force_x11: bool,
     window_ignore_dpi: bool,
@@ -66,11 +68,11 @@ pub struct BasaltOptions {
 impl Default for BasaltOptions {
     fn default() -> Self {
         Self {
-            require_instance_extensions: InstanceExtensions {
+            require_instance_extensions: vk::InstanceExtensions {
                 khr_surface: true,
-                ..InstanceExtensions::empty()
+                ..vk::InstanceExtensions::empty()
             },
-            prefer_instance_extensions: InstanceExtensions {
+            prefer_instance_extensions: vk::InstanceExtensions {
                 khr_xlib_surface: true,
                 khr_xcb_surface: true,
                 khr_wayland_surface: true,
@@ -82,23 +84,23 @@ impl Default for BasaltOptions {
                 khr_get_surface_capabilities2: true,
                 ext_surface_maintenance1: true,
                 ext_swapchain_colorspace: true,
-                ..InstanceExtensions::empty()
+                ..vk::InstanceExtensions::empty()
             },
             portability_subset: false,
             prefer_integrated_gpu: true,
-            require_device_extensions: DeviceExtensions::empty(),
-            prefer_device_extensions: DeviceExtensions {
+            require_device_extensions: vk::DeviceExtensions::empty(),
+            prefer_device_extensions: vk::DeviceExtensions {
                 ext_swapchain_maintenance1: true,
-                ..DeviceExtensions::empty()
+                ..vk::DeviceExtensions::empty()
             },
-            require_device_features: DeviceFeatures {
+            require_device_features: vk::DeviceFeatures {
                 descriptor_indexing: true,
                 shader_sampled_image_array_non_uniform_indexing: true,
                 runtime_descriptor_array: true,
                 descriptor_binding_variable_descriptor_count: true,
-                ..DeviceFeatures::empty()
+                ..vk::DeviceFeatures::empty()
             },
-            prefer_device_features: DeviceFeatures::empty(),
+            prefer_device_features: vk::DeviceFeatures::empty(),
             winit_force_x11: false,
             window_ignore_dpi: false,
             window_default_scale: 1.0,
@@ -123,13 +125,13 @@ impl BasaltOptions {
     ///
     /// ***Note:** This will cause an error if an extension is not supported. If this is not desired
     /// use the `prefer_instance_extensions` method instead.*
-    pub fn require_instance_extensions(mut self, extensions: InstanceExtensions) -> Self {
+    pub fn require_instance_extensions(mut self, extensions: vk::InstanceExtensions) -> Self {
         self.require_instance_extensions |= extensions;
         self
     }
 
     /// Add preferred instance extensions
-    pub fn prefer_instance_extensions(mut self, extensions: InstanceExtensions) -> Self {
+    pub fn prefer_instance_extensions(mut self, extensions: vk::InstanceExtensions) -> Self {
         self.prefer_instance_extensions |= extensions;
         self
     }
@@ -156,13 +158,13 @@ impl BasaltOptions {
     ///
     /// ***Note:** This will cause an error if an extension is not supported. If this is not desired
     /// use the `prefer_device_extensions` method instead.*
-    pub fn require_device_extensions(mut self, extensions: DeviceExtensions) -> Self {
+    pub fn require_device_extensions(mut self, extensions: vk::DeviceExtensions) -> Self {
         self.require_device_extensions |= extensions;
         self
     }
 
     /// Add preferred device extensions
-    pub fn prefer_device_extensions(mut self, extensions: DeviceExtensions) -> Self {
+    pub fn prefer_device_extensions(mut self, extensions: vk::DeviceExtensions) -> Self {
         self.prefer_device_extensions |= extensions;
         self
     }
@@ -171,13 +173,13 @@ impl BasaltOptions {
     ///
     /// ***Note:** This will cause an error if an feature is not supported. If this is not desired
     /// use the `prefer_device_features` method instead.*
-    pub fn require_device_features(mut self, features: DeviceFeatures) -> Self {
+    pub fn require_device_features(mut self, features: vk::DeviceFeatures) -> Self {
         self.require_device_features |= features;
         self
     }
 
     /// Add preferred device features
-    pub fn prefer_device_features(mut self, features: DeviceFeatures) -> Self {
+    pub fn prefer_device_features(mut self, features: vk::DeviceFeatures) -> Self {
         self.prefer_device_features |= features;
         self
     }
@@ -269,15 +271,15 @@ struct BasaltConfig {
 /// - This is expected to be kept alive for the lifetime of the application.
 /// - There should only ever be one instance of this struct.
 pub struct Basalt {
-    device: Arc<Device>,
+    device: Arc<vk::Device>,
     device_resources: Arc<vk::Resources>,
-    graphics_queue: Arc<device::Queue>,
-    transfer_queue: Arc<device::Queue>,
-    compute_queue: Arc<device::Queue>,
-    secondary_graphics_queue: Option<Arc<device::Queue>>,
-    secondary_transfer_queue: Option<Arc<device::Queue>>,
-    secondary_compute_queue: Option<Arc<device::Queue>>,
-    instance: Arc<Instance>,
+    graphics_queue: Arc<vk::Queue>,
+    transfer_queue: Arc<vk::Queue>,
+    compute_queue: Arc<vk::Queue>,
+    secondary_graphics_queue: Option<Arc<vk::Queue>>,
+    secondary_transfer_queue: Option<Arc<vk::Queue>>,
+    secondary_compute_queue: Option<Arc<vk::Queue>>,
+    instance: Arc<vk::Instance>,
     interface: Arc<Interface>,
     input: Input,
     interval: Arc<Interval>,
@@ -291,10 +293,10 @@ impl Basalt {
     /// Begin initializing Basalt, this thread will be taken for window event polling and the
     /// function provided in `result_fn` will be executed after Basalt initialization has
     /// completed or errored.
-    pub fn initialize<F: FnMut(Result<Arc<Self>, String>) + Send + 'static>(
-        options: BasaltOptions,
-        mut result_fn: F,
-    ) {
+    pub fn initialize<F>(options: BasaltOptions, result_fn: F)
+    where
+        F: FnOnce(Result<Arc<Self>, InitializeError>) + Send + 'static,
+    {
         let BasaltOptions {
             portability_subset,
             prefer_integrated_gpu,
@@ -314,9 +316,9 @@ impl Basalt {
             binary_fonts,
         } = options;
 
-        let vulkan_library = match VulkanLibrary::new() {
+        let vulkan_library = match vk::VulkanLibrary::new() {
             Ok(ok) => ok,
-            Err(e) => return result_fn(Err(format!("Failed to load vulkan library: {}", e))),
+            Err(e) => return result_fn(Err(InitializeError::LoadVulkanLibrary(e))),
         };
 
         let instance_extensions = vulkan_library
@@ -324,88 +326,123 @@ impl Basalt {
             .intersection(&prefer_instance_extensions)
             .union(&require_instance_extensions);
 
-        let mut instance_create_flags = InstanceCreateFlags::empty();
+        let mut instance_create_flags = vk::InstanceCreateFlags::empty();
 
         if portability_subset {
-            instance_create_flags |= InstanceCreateFlags::ENUMERATE_PORTABILITY;
+            instance_create_flags |= vk::InstanceCreateFlags::ENUMERATE_PORTABILITY;
         }
 
-        let instance = match Instance::new(
+        let instance = match vk::Instance::new(
             vulkan_library,
-            InstanceCreateInfo {
+            vk::InstanceCreateInfo {
                 flags: instance_create_flags,
                 enabled_extensions: instance_extensions,
                 engine_name: Some(String::from("Basalt")),
-                engine_version: Version {
+                engine_version: vk::Version {
                     major: 0,
                     minor: 21,
                     patch: 0,
                 },
-                ..InstanceCreateInfo::default()
+                ..Default::default()
             },
         ) {
             Ok(ok) => ok,
-            Err(e) => return result_fn(Err(format!("Failed to create instance: {}", e))),
+            Err(e) => return result_fn(Err(InitializeError::CreateInstance(e))),
         };
 
-        if instance.api_version() < Version::V1_2 {
-            return result_fn(Err(String::from(
-                "Basalt requires vulkan version 1.2 or above",
-            )));
+        if instance.api_version() < vk::Version::V1_2 {
+            return result_fn(Err(InitializeError::IncompatibleVulkan));
         }
 
         WindowManager::run(winit_force_x11, move |window_manager| {
             let mut physical_devices = match instance.enumerate_physical_devices() {
                 Ok(ok) => ok.collect::<Vec<_>>(),
                 Err(e) => {
-                    return result_fn(Err(format!("Failed to enumerate physical devices: {}", e)));
+                    return result_fn(Err(InitializeError::EnumerateDevices(e)));
                 },
             };
 
             if prefer_integrated_gpu {
                 physical_devices.sort_by_key(|dev| {
                     match dev.properties().device_type {
-                        PhysicalDeviceType::DiscreteGpu => 4,
-                        PhysicalDeviceType::IntegratedGpu => 5,
-                        PhysicalDeviceType::VirtualGpu => 3,
-                        PhysicalDeviceType::Other => 2,
-                        PhysicalDeviceType::Cpu => 1,
+                        vk::PhysicalDeviceType::DiscreteGpu => 4,
+                        vk::PhysicalDeviceType::IntegratedGpu => 5,
+                        vk::PhysicalDeviceType::VirtualGpu => 3,
+                        vk::PhysicalDeviceType::Other => 2,
+                        vk::PhysicalDeviceType::Cpu => 1,
                         _ => 0,
                     }
                 });
             } else {
                 physical_devices.sort_by_key(|dev| {
                     match dev.properties().device_type {
-                        PhysicalDeviceType::DiscreteGpu => 5,
-                        PhysicalDeviceType::IntegratedGpu => 4,
-                        PhysicalDeviceType::VirtualGpu => 3,
-                        PhysicalDeviceType::Other => 2,
-                        PhysicalDeviceType::Cpu => 1,
+                        vk::PhysicalDeviceType::DiscreteGpu => 5,
+                        vk::PhysicalDeviceType::IntegratedGpu => 4,
+                        vk::PhysicalDeviceType::VirtualGpu => 3,
+                        vk::PhysicalDeviceType::Other => 2,
+                        vk::PhysicalDeviceType::Cpu => 1,
                         _ => 0,
                     }
                 });
             }
 
+            physical_devices.retain(|physical_device| {
+                if physical_device.api_version() < vk::Version::V1_2 {
+                    println!(
+                        "[Basalt]: Unable to use physical device, {}: api version < 1.2.",
+                        physical_device.properties().device_name
+                    );
+                    return false;
+                }
+
+                if physical_device
+                    .supported_features()
+                    .contains(&require_device_features)
+                {
+                    println!(
+                        "[Basalt]: Unable to use physical device, {}: missing required device \
+                         features.",
+                        physical_device.properties().device_name
+                    );
+                    return false;
+                }
+
+                if physical_device
+                    .supported_extensions()
+                    .contains(&require_device_extensions)
+                {
+                    println!(
+                        "[Basalt]: Unable to use physical device, {}: missing required device \
+                         extentsions.",
+                        physical_device.properties().device_name
+                    );
+                    return false;
+                }
+
+                true
+            });
+
             let physical_device = match physical_devices.pop() {
                 Some(some) => some,
-                None => return result_fn(Err(String::from("No suitable device found."))),
+                None => return result_fn(Err(InitializeError::NoSuitableDevice)),
             };
 
-            let mut available_queue_families: BTreeMap<u32, (QueueFlags, u32)> = BTreeMap::new();
+            let mut available_queue_families: BTreeMap<u32, (vk::QueueFlags, u32)> =
+                BTreeMap::new();
             let mut graphics_queue_families: Vec<u32> = Vec::new();
             let mut compute_queue_families: Vec<u32> = Vec::new();
             let mut transfer_queue_families: Vec<u32> = Vec::new();
 
             for (i, properties) in physical_device.queue_family_properties().iter().enumerate() {
-                if properties.queue_flags.contains(QueueFlags::GRAPHICS) {
+                if properties.queue_flags.contains(vk::QueueFlags::GRAPHICS) {
                     graphics_queue_families.push(i as u32);
                 }
 
-                if properties.queue_flags.contains(QueueFlags::COMPUTE) {
+                if properties.queue_flags.contains(vk::QueueFlags::COMPUTE) {
                     compute_queue_families.push(i as u32);
                 }
 
-                if properties.queue_flags.contains(QueueFlags::TRANSFER) {
+                if properties.queue_flags.contains(vk::QueueFlags::TRANSFER) {
                     transfer_queue_families.push(i as u32);
                 }
 
@@ -416,39 +453,39 @@ impl Basalt {
             graphics_queue_families.sort_by_cached_key(|index| {
                 let flags = available_queue_families.get(index).unwrap().0;
                 let mut weight: u8 = 0;
-                weight += flags.contains(QueueFlags::COMPUTE) as u8;
-                weight += flags.contains(QueueFlags::PROTECTED) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_DECODE) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_ENCODE) as u8;
-                weight += flags.contains(QueueFlags::OPTICAL_FLOW) as u8;
+                weight += flags.contains(vk::QueueFlags::COMPUTE) as u8;
+                weight += flags.contains(vk::QueueFlags::PROTECTED) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_DECODE) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_ENCODE) as u8;
+                weight += flags.contains(vk::QueueFlags::OPTICAL_FLOW) as u8;
                 weight
             });
 
             compute_queue_families.sort_by_cached_key(|index| {
                 let flags = available_queue_families.get(index).unwrap().0;
                 let mut weight: u8 = 0;
-                weight += flags.contains(QueueFlags::GRAPHICS) as u8;
-                weight += flags.contains(QueueFlags::PROTECTED) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_DECODE) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_ENCODE) as u8;
-                weight += flags.contains(QueueFlags::OPTICAL_FLOW) as u8;
+                weight += flags.contains(vk::QueueFlags::GRAPHICS) as u8;
+                weight += flags.contains(vk::QueueFlags::PROTECTED) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_DECODE) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_ENCODE) as u8;
+                weight += flags.contains(vk::QueueFlags::OPTICAL_FLOW) as u8;
                 weight
             });
 
             transfer_queue_families.sort_by_cached_key(|index| {
                 let flags = available_queue_families.get(index).unwrap().0;
                 let mut weight: u8 = 0;
-                weight += flags.contains(QueueFlags::GRAPHICS) as u8;
-                weight += flags.contains(QueueFlags::COMPUTE) as u8;
-                weight += flags.contains(QueueFlags::PROTECTED) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_DECODE) as u8;
-                weight += flags.contains(QueueFlags::VIDEO_ENCODE) as u8;
-                weight += flags.contains(QueueFlags::OPTICAL_FLOW) as u8;
+                weight += flags.contains(vk::QueueFlags::GRAPHICS) as u8;
+                weight += flags.contains(vk::QueueFlags::COMPUTE) as u8;
+                weight += flags.contains(vk::QueueFlags::PROTECTED) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_DECODE) as u8;
+                weight += flags.contains(vk::QueueFlags::VIDEO_ENCODE) as u8;
+                weight += flags.contains(vk::QueueFlags::OPTICAL_FLOW) as u8;
                 weight
             });
 
             let select_queue =
-                |indexes: &Vec<u32>, queue_families: &mut BTreeMap<u32, (QueueFlags, u32)>| {
+                |indexes: &Vec<u32>, queue_families: &mut BTreeMap<u32, (vk::QueueFlags, u32)>| {
                     let mut selected_index = None;
 
                     for index in indexes.iter() {
@@ -516,7 +553,7 @@ impl Basalt {
             let mut queue_map: Vec<(usize, usize)> = Vec::new();
             let mut queue_count = 0;
 
-            let queue_request: Vec<QueueCreateInfo> = family_map
+            let queue_request: Vec<vk::QueueCreateInfo> = family_map
                 .into_iter()
                 .map(|(family_index, members)| {
                     let mut priorites = Vec::with_capacity(members.len());
@@ -527,7 +564,7 @@ impl Basalt {
                         priorites.push(priority);
                     }
 
-                    QueueCreateInfo {
+                    vk::QueueCreateInfo {
                         queues: priorites,
                         queue_family_index: family_index,
                         ..Default::default()
@@ -545,27 +582,22 @@ impl Basalt {
                 .intersection(&prefer_device_features)
                 .union(&require_device_features);
 
-            let (device, queues) = match Device::new(
+            let (device, queues) = match vk::Device::new(
                 physical_device,
-                DeviceCreateInfo {
+                vk::DeviceCreateInfo {
                     enabled_extensions: device_extensions,
                     enabled_features: device_features,
                     queue_create_infos: queue_request,
-                    ..DeviceCreateInfo::default()
+                    ..Default::default()
                 },
             ) {
                 Ok(ok) => ok,
-                Err(e) => return result_fn(Err(format!("Failed to create device: {:?}", e))),
+                Err(e) => return result_fn(Err(InitializeError::CreateDevice(e))),
             };
 
-            if queues.len() != queue_map.len() {
-                return result_fn(Err(String::from(
-                    "Returned queues length != expected length",
-                )));
-            }
+            assert_eq!(queues.len(), queue_map.len());
 
-            let mut queues: Vec<Option<Arc<device::Queue>>> =
-                queues.into_iter().map(Some).collect();
+            let mut queues: Vec<Option<Arc<vk::Queue>>> = queues.into_iter().map(Some).collect();
             let mut graphics_queue = None;
             let mut secondary_graphics_queue = None;
             let mut compute_queue = None;
@@ -688,60 +720,62 @@ impl Basalt {
     }
 
     /// Obtain a copy of `Arc<Instance>`
-    pub fn instance(&self) -> Arc<Instance> {
+    pub fn instance(&self) -> Arc<vk::Instance> {
         self.instance.clone()
     }
 
     /// Obtain a reference of `Arc<Instance>`
-    pub fn instance_ref(&self) -> &Arc<Instance> {
+    pub fn instance_ref(&self) -> &Arc<vk::Instance> {
         &self.instance
     }
 
     /// Obtain a copy of `Arc<PhysicalDevice>`
-    pub fn physical_device(&self) -> Arc<PhysicalDevice> {
+    pub fn physical_device(&self) -> Arc<vk::PhysicalDevice> {
         self.device.physical_device().clone()
     }
 
     /// Obtain a reference of `Arc<PhysicalDevice>`
-    pub fn physical_device_ref(&self) -> &Arc<PhysicalDevice> {
+    pub fn physical_device_ref(&self) -> &Arc<vk::PhysicalDevice> {
         self.device.physical_device()
     }
 
     /// Obtain a copy of `Arc<Devcie>`
-    pub fn device(&self) -> Arc<Device> {
+    pub fn device(&self) -> Arc<vk::Device> {
         self.device.clone()
     }
 
     /// Obtain a refernce of `Arc<Device>`
-    pub fn device_ref(&self) -> &Arc<Device> {
+    pub fn device_ref(&self) -> &Arc<vk::Device> {
         &self.device
     }
 
+    /// Obtain a copy of `Arc<Resources>`.
     pub fn device_resources(&self) -> Arc<vk::Resources> {
         self.device_resources.clone()
     }
 
+    /// Obtain a reference of `Arc<Resources>`.
     pub fn device_resources_ref(&self) -> &Arc<vk::Resources> {
         &self.device_resources
     }
 
     /// Obtain a copy of the `Arc<Queue>` assigned for graphics operations.
-    pub fn graphics_queue(&self) -> Arc<device::Queue> {
+    pub fn graphics_queue(&self) -> Arc<vk::Queue> {
         self.graphics_queue.clone()
     }
 
     /// Obtain a reference of the `Arc<Queue>` assigned for graphics operations.
-    pub fn graphics_queue_ref(&self) -> &Arc<device::Queue> {
+    pub fn graphics_queue_ref(&self) -> &Arc<vk::Queue> {
         &self.graphics_queue
     }
 
     /// Obtain a copy of the `Arc<Queue>` assigned for secondary graphics operations.
-    pub fn secondary_graphics_queue(&self) -> Option<Arc<device::Queue>> {
+    pub fn secondary_graphics_queue(&self) -> Option<Arc<vk::Queue>> {
         self.secondary_graphics_queue.clone()
     }
 
     /// Obtain a reference of the `Arc<Queue>` assigned for secondary graphics operations.
-    pub fn secondary_graphics_queue_ref(&self) -> Option<&Arc<device::Queue>> {
+    pub fn secondary_graphics_queue_ref(&self) -> Option<&Arc<vk::Queue>> {
         self.secondary_graphics_queue.as_ref()
     }
 
@@ -750,7 +784,7 @@ impl Basalt {
     /// # Notes:
     /// - This queue may be the same as the graphics queue in cases where the device only
     /// has a single queue present.
-    pub fn compute_queue(&self) -> Arc<device::Queue> {
+    pub fn compute_queue(&self) -> Arc<vk::Queue> {
         self.compute_queue.clone()
     }
 
@@ -759,17 +793,17 @@ impl Basalt {
     /// # Notes:
     /// - This queue may be the same as the graphics queue in cases where the device only
     /// has a single queue present.
-    pub fn compute_queue_ref(&self) -> &Arc<device::Queue> {
+    pub fn compute_queue_ref(&self) -> &Arc<vk::Queue> {
         &self.compute_queue
     }
 
     /// Obtain a copy of the `Arc<Queue>` assigned for secondary compute operations.
-    pub fn secondary_compute_queue(&self) -> Option<Arc<device::Queue>> {
+    pub fn secondary_compute_queue(&self) -> Option<Arc<vk::Queue>> {
         self.secondary_compute_queue.clone()
     }
 
     /// Obtain a reference of the `Arc<Queue>` assigned for secondary compute operations.
-    pub fn secondary_compute_queue_ref(&self) -> Option<&Arc<device::Queue>> {
+    pub fn secondary_compute_queue_ref(&self) -> Option<&Arc<vk::Queue>> {
         self.secondary_compute_queue.as_ref()
     }
 
@@ -779,7 +813,7 @@ impl Basalt {
     /// - This queue may be the same as the compute queue in cases where the device only
     /// has two queues present. In cases where there is only one queue the graphics, compute,
     /// and transfer queues will all be the same queue.
-    pub fn transfer_queue(&self) -> Arc<device::Queue> {
+    pub fn transfer_queue(&self) -> Arc<vk::Queue> {
         self.transfer_queue.clone()
     }
 
@@ -789,17 +823,17 @@ impl Basalt {
     /// - This queue may be the same as the compute queue in cases where the device only
     /// has two queues present. In cases where there is only one queue the graphics, compute,
     /// and transfer queues will all be the same queue.
-    pub fn transfer_queue_ref(&self) -> &Arc<device::Queue> {
+    pub fn transfer_queue_ref(&self) -> &Arc<vk::Queue> {
         &self.transfer_queue
     }
 
     /// Obtain a copy of the `Arc<Queue>` assigned for secondary transfers.
-    pub fn secondary_transfer_queue(&self) -> Option<Arc<device::Queue>> {
+    pub fn secondary_transfer_queue(&self) -> Option<Arc<vk::Queue>> {
         self.secondary_transfer_queue.clone()
     }
 
     /// Obtain a reference of the `Arc<Queue>` assigned for secondary transfers.
-    pub fn secondary_transfer_queue_ref(&self) -> Option<&Arc<device::Queue>> {
+    pub fn secondary_transfer_queue_ref(&self) -> Option<&Arc<vk::Queue>> {
         self.secondary_transfer_queue.as_ref()
     }
 
@@ -817,7 +851,47 @@ impl Basalt {
 
 impl std::fmt::Debug for Basalt {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Basalt").finish()
+        f.debug_struct("Basalt").finish_non_exhaustive()
+    }
+}
+
+/// An error occurred during the `Basalt`'s initialization.
+#[derive(Debug)]
+pub enum InitializeError {
+    /// Failed to load the `VulkanLibrary`.
+    LoadVulkanLibrary(vk::LoadingError),
+    /// Failed to create the `Instance`.
+    CreateInstance(vk::Validated<vk::VulkanError>),
+    /// The `Instance`'s vulkan version is less than  1.2.
+    IncompatibleVulkan,
+    /// Failed to enumerate the physical devices.
+    EnumerateDevices(vk::VulkanError),
+    /// There are no suitable devices.
+    NoSuitableDevice,
+    /// Failed to create the `Device`
+    CreateDevice(vk::Validated<vk::VulkanError>),
+}
+
+impl Display for InitializeError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), fmt::Error> {
+        match self {
+            Self::LoadVulkanLibrary(e) => {
+                write!(f, "Failed to load the vulkan library: {}", e)
+            },
+            Self::CreateInstance(e) => {
+                write!(f, "Failed to create the instance: {}", e)
+            },
+            Self::IncompatibleVulkan => {
+                f.write_str("The vulkan instance's version is less than 1.2.")
+            },
+            Self::EnumerateDevices(e) => {
+                write!(f, "Failed to enumerate the physical devices: {}", e)
+            },
+            Self::NoSuitableDevice => f.write_str("No suitable device was found."),
+            Self::CreateDevice(e) => {
+                write!(f, "Failed to create the device: {}", e)
+            },
+        }
     }
 }